fix: update symbol regex
The new symbol regex definition can be summarized as follows:

Symbols cannot start with uppercase letters, because these are reserved. Moreover, any word between quotes is a symbol.
More in detail:
1) either start with [a-z_], followed by at least one [a-zA-Z0-9_-], and by one [a-zA-Z0-9_] (i.e. hyphens only in between)
2) or, start with [a-z_] and follows with any sequence of [a-zA-Z0-9_] (no hyphens)
3) or, any sequence of ASCII printable characters (i.e. going from ' ' to '~'), except '"'.

The previous symbol regex was allowing symbols that end with hyphens, e.g. 'a-'. The new regex does not allow this.
As first symbol, now also '_' is allowed. Moreover, in double quotes, all printable ASCII characters except '"' are allowed.